backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
transformers==4.45.2
torch==2.9.0
sentence-transformers==2.2.2
Pillow==11.0.0
numpy==2.1.3
pandas==2.2.3
scikit-learn==1.5.2
python-multipart==0.0.6
pydantic==2.9.2
openpyxl==3.1.2
tokenizers==0.20.1

